Removed nhccoco folder and moved new stuff into root
