+++ custom_components/nhc2/coco_device_class.py
@@ -1,11 +1,14 @@
 from enum import Enum
 
-
 class CoCoDeviceClass(Enum):
     SWITCHES = 'switches'
     LIGHTS = 'lights'
     SHUTTERS = 'shutters'
+    COVERS = 'covers'
+    GATE = 'gate'
     FANS = 'fans'
     SWITCHED_FANS = 'switched-fans'
     THERMOSTATS = 'thermostats'
+    ENERGYMETERS = 'energymeters'
+    ACCESSCONTROL = 'accesscontrol'
     GENERIC = 'generic'
