--- custom_components/nhc2/nhccoco/__init__.py
@@ -1,21 +1,0 @@
-from .coco import CoCo
-from .coco_entity import CoCoEntity
-from .coco_light import CoCoLight
-from .coco_switch import CoCoSwitch
-from .coco_fan import CoCoFan
-from .coco_climate import CoCoThermostat
-from .coco_energy import CoCoEnergyMeter
-from .coco_cover import CoCoCover
-from .coco_accesscontrol import CoCoAccessControl
-from .coco_device_class import CoCoDeviceClass
-
-__all__ = ["CoCo",
-           "CoCoEntity",
-           "CoCoLight",
-           "CoCoSwitch",
-           "CoCoFan",
-           "CoCoThermostat",
-           "CoCoEnergyMeter",
-           "CoCoCover",
-           "CoCoAccessControl",
-           "CoCoDeviceClass"]
