--- custom_components/nhc2/nhccoco/coco_device_class.py
@@ -1,13 +1,0 @@
-from enum import Enum
-
-class CoCoDeviceClass(Enum):
-    SWITCHES = 'switches'
-    LIGHTS = 'lights'
-    COVERS = 'covers'
-    GATE = 'gate'
-    FANS = 'fans'
-    SWITCHED_FANS = 'switched-fans'
-    THERMOSTATS = 'thermostats'
-    ENERGYMETERS = 'energymeters'
-    ACCESSCONTROL = 'accesscontrol'
-    GENERIC = 'generic'
